svetovit/requirements.txt
python-gitlab==5.3.1
